refactor: 상품 삭제 시 공동구매 존재하는 경우 soft delete

+++ product/src/main/java/store/_0982/product/presentation/dto/ProductRegisterRequest.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 public record ProductRegisterRequest(
         String name,
-        int price,
+        Long price,
         ProductCategory category,
         String description,
         int stock,
